+++ projects/npm-tools/packages/npm-scripts/src/scripts/lint.js
@@ -9,6 +9,7 @@ const path = require('path');
 
 const isJSP = require('../jsp/isJSP');
 const lintJSP = require('../jsp/lintJSP');
+const color = require('../utils/color');
 const getMergedConfig = require('../utils/getMergedConfig');
 const getPaths = require('../utils/getPaths');
 const log = require('../utils/log');
@@ -162,29 +163,6 @@ async function lint(options = {}) {
 		throw new SpawnError();
 	}
 }
-
-function color(name) {
-	if (process.stdout.isTTY) {
-		return (
-			{
-				BOLD: '\x1b[1m',
-				RED: '\x1b[31m',
-				RESET: '\x1b[0m',
-				UNDERLINE: '\x1b[4m',
-				YELLOW: '\x1b[33m',
-			}[name] || ''
-		);
-	}
-	else {
-		return '';
-	}
-}
-
-color.BOLD = color('BOLD');
-color.RED = color('RED');
-color.RESET = color('RESET');
-color.UNDERLINE = color('UNDERLINE');
-color.YELLOW = color('YELLOW');
 
 /**
  * @see https://codepoints.net/U+2716?lang=en
